cleaned a little bit of code up and moved ClusterComponentInfo into the graphcompponents folder
